refactor(node/sync): extract namespace_sync submodule from the manager god-file (increment 1)
First increment of splitting `sync/manager/mod.rs` (4720 LOC). Moves the
namespace sync `impl SyncManager` block (~1290 lines) into `namespace_sync.rs`
— behavior-preserving, pure code movement (mod.rs: -1290 block, + wiring).

Moved (governance catch-up/backfill + namespace / open-subgroup join):
- request_governance_catchup_from_peer, backfill_governance_for_pending_deltas
- handle_namespace_backfill_request, handle_namespace_join_request,
  handle_open_subgroup_join_request
- initiate_namespace_join, initiate_open_subgroup_join, sync_namespace_from_peer
- drain_governance_pending_after_sync, collect_namespace_governance_ops
- the resolve_namespace_id / should_backfill_governance free helpers

Wiring: methods called from the stream dispatcher / `SyncDriverDispatch` trait
in mod.rs are `pub(super)`; methods that stay in mod.rs remain reachable from
the moved code via ancestor privacy (child module sees ancestor private items),
so no visibility churn there. `super::{config,peers,stream}` paths (sync-level
siblings) rewritten to `crate::sync::*`; `namespace_join` to `super::`. The two
free helpers are `pub(super)` and re-exported under `#[cfg(test)]` for the
tests submodule's `super::super::` path. Explicit imports, no `use super::*`.

Validated: node lib tests (333) and sync_sim (314) pass.
